Adds bootstrapping option
When bootstrapping, we configure the cluster address to only include the
nodes that have previously been setup.

Setting bootstrapping to true results in the following cluster config:

e.g.

  db01 gcomm://
  db02 gcomm://db01
  db03 gcomm://db01,db02

This allows the cluster to be easily bootstrapped by just running chef
on the nodes in order.

After bootstrapping the cluster, we want to include all addresses for
all nodes. If we don't do this, when db01 is restarted it will not
rejoin the cluster, as its cluster address contains no nodes. This
results in a split-brain scenario.

Setting bootstrapping to false results in the following cluster config:

e.g.

  db01 gcomm://db01,db02,db03
  db02 gcomm://db01,db02,db03
  db03 gcomm://db01,db02,db03

This ensures that when nodes are restarted, they will rejoin the
cluster. The bootstrapping configuration will then remain unchanged as
long as the cluster is around.