Adds bind_address, dev tooling
Instead of getting fancy with finding the node ip (which, as it turns
out, was never being called because the percona bind address doesn't
default to nil), we now require a bind_address. This is a breaking
change, and so we will increment the major version when this is
released.

Also, some development tooling has been added (foodcritic, rubocop, and
chefspec).